ai_menu_reader/requirements.txt
fastapi
uvicorn
paddleocr
paddlepaddle
opencv-python
ultralytics
google-generativeai
numpy
pillow